backend/assets/aéncientrequirements .txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.6
django-allauth==0.44.0
django-auto-prefetching==0.1.9
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-extensions==3.1.1
django-import-export==2.5.0
django-rest-auth==0.9.5
django-rest-multiple-models==2.1.3
django-restql==0.11.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-writable-nested==0.6.2
et-xmlfile==1.1.0
idna==2.10
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.1
pyPEG2==2.15.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
